+++ dex/farm/src/external_interaction.rs
@@ -16,6 +16,7 @@ pub trait ExternalInteractionsModule:
     + pausable::PausableModule
     + permissions_module::PermissionsModule
     + permissions_hub_module::PermissionsHubModule
+    + original_owner_helper::OriginalOwnerHelperModule
     + sc_whitelist_module::SCWhitelistModule
     + events::EventsModule
     + multiversx_sc_modules::default_issue_callbacks::DefaultIssueCallbacksModule
