+++ dex/farm/src/lib.rs
@@ -31,6 +31,7 @@ pub trait Farm:
     + pausable::PausableModule
     + permissions_module::PermissionsModule
     + permissions_hub_module::PermissionsHubModule
+    + original_owner_helper::OriginalOwnerHelperModule
     + sc_whitelist_module::SCWhitelistModule
     + events::EventsModule
     + multiversx_sc_modules::default_issue_callbacks::DefaultIssueCallbacksModule
